:recycle: (pc-15004) move useEduconnectLogin to identityCheck/api
