+++ src/features/identityCheck/pages/identification/IdentityCheckEduConnectForm.web.tsx
@@ -3,10 +3,10 @@ import React from 'react'
 import { ErrorBoundary } from 'react-error-boundary'
 import styled from 'styled-components/native'
 
+import { useEduConnectLogin } from 'features/identityCheck/api/useEduConnectLogin'
 import { ErrorTrigger } from 'features/identityCheck/atoms/ErrorTrigger'
 import { PageWithHeader } from 'features/identityCheck/components/layout/PageWithHeader'
 import { EduConnectErrorBoundary } from 'features/identityCheck/errors/eduConnect/EduConnectErrorBoundary'
-import { useEduConnectLogin } from 'features/identityCheck/utils/useEduConnectLogin'
 import { ButtonPrimary } from 'ui/components/buttons/ButtonPrimary'
 import { BicolorIdCardWithMagnifyingGlass } from 'ui/svg/icons/BicolorIdCardWithMagnifyingGlass'
 import { ExternalSite } from 'ui/svg/icons/ExternalSite'
